refactor(librarian): Move generate_e2e_test.go to internal/librarian
Moved generate_e2e_test.go from the root directory to internal/librarian/
to co-locate it with the generate command's core logic.

This change also involved:
- Extracting the common test helper function runGit into internal/librarian/test_helpers.go
  to resolve name collisions.
- Updating relative paths to testdata in generate_e2e_test.go.
- Resolving unused imports and formatting issues.
